+++ src/data-hash-crc32.js
@@ -41,7 +41,7 @@ export const CRC32_TABLE = [
  * @see {@link https://rosettacode.org/wiki/CRC-32|CRC-32}
  * @see {@link https://en.wikipedia.org/wiki/Computation_of_cyclic_redundancy_checks|Computation of cyclic redundancy checks}
  * @example <caption>CRC32.of(...)</caption>
- * import CRC32 from 'uttori-data-tools/data-hash-crc32';
+ * import CRC32 from '@uttori/data-tools/data-hash-crc32';
  * CRC32.of('The quick brown fox jumps over the lazy dog');
  * ➜ '414FA339'
  * @param {number[]|ArrayBuffer|Buffer|DataBuffer|Int8Array|Int16Array|Int32Array|number|string|Uint8Array|Uint16Array|Uint32Array} data The data to process.